use upstream CrumbRequester if available
